Percent chance, metafield for alf in lieu of REST API to get offer details, cleanup offer detail naming, allow deleting SKU from offer

+++ src/app/api/manifest/route.ts
@@ -3,8 +3,15 @@ import db from '@/lib/db'
 import updateOfferManifests from '@/app/api/manifest/updateOfferManifest'
 import queryOffer from '@/app/api/manifest/queryOffer'
 import { z } from 'zod'
-import { V3OfferListItem } from '@/app/api/manifest/models'
+import { V3Offer, V3OfferListItem } from '@/app/api/manifest/models'
 import { getSession } from '@/lib/session'
+import shopifyWriteVariantMetafield from '@/lib/shopifyWriteVariantMetafield'
+import {
+  shopifyGetProductDataByVariantId,
+  shopifyGetProductDataByVariantIds,
+} from '@/lib/shopifyGetProductData'
+import { parse } from 'date-fns'
+import shopifyWriteProductMetafield from '@/lib/shopifyWriteProductMetafield'
 
 export async function POST(req: NextRequest): Promise<NextResponse> {
   const session = await getSession()
@@ -19,8 +26,27 @@ export async function POST(req: NextRequest): Promise<NextResponse> {
     try {
       const SELECT =
         'select offer_id, offer_name, offer_variant_id from v3_offer order by offer_id desc'
+      const offerListItems: V3OfferListItem[] = []
+      const raw_v3_offer_rows: {
+        offer_id: number
+        offer_name: string
+        offer_variant_id: string
+      }[] = await db.query(SELECT)
 
-      const existingOfferList: V3OfferListItem[] = await db.query(SELECT)
+      const offerProductData = await shopifyGetProductDataByVariantIds(
+        raw_v3_offer_rows.map((r) => r.offer_variant_id),
+      )
+
+      for (const offer of raw_v3_offer_rows) {
+        offerListItems.push({
+          offer_id: offer.offer_id,
+          offer_name: offer.offer_name,
+          offerProductData: {
+            ...offerProductData[offer.offer_variant_id],
+            variantId: offer.offer_variant_id,
+          },
+        })
+      }
 
       switch (o.action) {
         case 'offer_create':
@@ -34,25 +60,44 @@ export async function POST(req: NextRequest): Promise<NextResponse> {
             'insert into v3_offer (offer_name, offer_variant_id) values (?, ?)',
             [parsed.offer_name, parsed.offer_variant_id],
           )
-          existingOfferList.push({
+          offerListItems.push({
             offer_id: 0,
             offer_name: o.name,
-            offer_variant_id: parsed.offer_variant_id,
+            offerProductData: {
+              ...(offerProductData[parsed.offer_variant_id] ??
+                (await shopifyGetProductDataByVariantId(
+                  parsed.offer_variant_id,
+                ))),
+              variantId: parsed.offer_variant_id,
+            },
           })
-          return NextResponse.json(existingOfferList)
+          return NextResponse.json(offerListItems)
 
         case 'offer_delete':
           await db.query(
             'delete from v3_offer_manifest where assignee_id is null and offer_id = ?',
             [o.offer_id],
           )
+          const remaining: any[] = await db.query(
+            'select count(*) c from v3_offer_manifest where offer_id = ?',
+            [o.offer_id],
+          )
+          if (remaining[0].c != 0) {
+            return NextResponse.json(offerListItems, {
+              status: 400,
+              statusText: 'Failed to delete offer due to allocated manifests',
+            })
+          }
           await db.query('delete from v3_offer where offer_id = ?', [
             o.offer_id,
           ])
+          return NextResponse.json(
+            offerListItems.filter((ol) => ol.offer_id != o.offer_id),
+          )
       }
 
       // now return the list
-      return NextResponse.json(await db.query(SELECT))
+      return NextResponse.json(offerListItems)
     } catch (err) {
       return NextResponse.json(err, { status: 400 })
     } finally {
@@ -60,21 +105,39 @@ export async function POST(req: NextRequest): Promise<NextResponse> {
     }
   }
 
+  // these have to do with an offer_id or name
+
   if (o.action === 'offer_get') {
     const params = z.object({ offer_id: z.coerce.number() }).parse(o)
     const offer = await queryOffer(params)
+    await maybeUpdateOfferMetafield(offer)
     return NextResponse.json(offer)
   }
 
   // More involved actions, these already wrap DB so don't put in the list above
   if (o.action === 'put_sku_to_offer') {
-    return NextResponse.json(
-      await updateOfferManifests(o.offer_id, o.sku_to_add),
-    )
+    const updatedOffer = await updateOfferManifests(o.offer_id, o.sku_to_add)
+    await maybeUpdateOfferMetafield(updatedOffer)
+    return NextResponse.json(updatedOffer)
   }
 
   if (o.action === 'shopify_webhook') {
   }
 
   return NextResponse.json(queryOffer({ offer_name: o.name }))
+}
+
+async function maybeUpdateOfferMetafield(updatedOffer: V3Offer | null) {
+  if (updatedOffer != null) {
+    await shopifyWriteProductMetafield(
+      updatedOffer.offerProductData.productId,
+      'offer_v3',
+      JSON.stringify(updatedOffer.manifestProductData, null, 2),
+    )
+    // await shopifyWriteVariantMetafield(
+    //   updatedOffer.offerProductData.variantId,
+    //   'product_data',
+    //   JSON.stringify(updatedOffer.manifestProductData),
+    // )
+  }
 }
